Orders controller

+++ controller/order.js
@@ -33,7 +33,11 @@ exports.store = async (req, res) => {
   });
 
   try {
-    await order.destroy({where: {transactionId}});
+    const userOrders = order.findAll({where: {transactionId}});
+    if (userOrders) {
+      await order.destroy({where: {transactionId}});
+    }
+
     const orderData = await order.bulkCreate(newOrders);
 
     return res.json({message: 'Ordes send', orderData});
